Former-commit-id: 1e085e32e9d729c1c3b8d48b67ef6cc483c291bf

+++ src/game_engine/actions/PointsChanger.java
@@ -6,10 +6,10 @@ import game_data.states.*;
 
 public abstract class PointsChanger implements Action {
 	
-	protected int pointsToGive;
+	protected double pointsToGive;
 	protected Sprite spriteAffected;
 	
-	public PointsChanger(int damage, Sprite sprite){
+	public PointsChanger(double damage, Sprite sprite){
 		pointsToGive = damage;
 		spriteAffected = sprite;
 	}
